style: renaming function to be action-oriented
